create v3/metadata module
refactor metadata into v2 and v3 modules

remove version signifiers from function names

fix tests
