+++ tests/v3/test_metadata/test_v2.py
@@ -10,17 +10,18 @@ if TYPE_CHECKING:
 import pytest
 
 from zarr.codecs import GzipCodec
-from zarr.metadata import ArrayV2Metadata, parse_zarr_format_v2
+from zarr.metadata import ArrayV2Metadata
+from zarr.metadata.v2 import parse_zarr_format
 
 
 def test_parse_zarr_format_valid() -> None:
-    assert parse_zarr_format_v2(2) == 2
+    assert parse_zarr_format(2) == 2
 
 
 @pytest.mark.parametrize("data", [None, 1, 3, 4, 5, "3"])
 def test_parse_zarr_format_invalid(data: Any) -> None:
     with pytest.raises(ValueError, match=f"Invalid value. Expected 2. Got {data}"):
-        parse_zarr_format_v2(data)
+        parse_zarr_format(data)
 
 
 @pytest.mark.parametrize("attributes", [None, {"foo": "bar"}])
